Refactoring the card.js file into smaller resuable components
